Refactor replay corpus facade
